v0.12.2 — strip TimeProvider to pure ABC, float volume
- Remove pandas_market_calendars dependency from TimeProvider (calendar
  logic moved to infrastructure/MarketSchedule in callers)
- Drop localize(), previous_trading_day(), last_completed_trading_day()
  from TimeProvider base; callers own their timezone handling
- Add combine(time) helper: returns tz-aware datetime for today at given time
- BarData.volume: int → float (IB returns float for some instruments)

+++ src/xtrading_models/bar.py
@@ -10,7 +10,7 @@ class BarData(BaseModel):
     high: float = 0.0
     low: float = 0.0
     close: float = 0.0
-    volume: int = 0
+    volume: float = 0
     is_close_bar: bool = False
     # average: float = 0.0
     # barCount: int = 0
